==> 01_Data_download.ipynb <==
 # Description of the notebook

 In this notebook I am downloaded shared data to my workspace.

 A dedicated API service was created for this workshop, available at: https://api-datalab.coderslab.com/api/v2. Additionally, documentation has been made available, which can be found here: [click](https://api-datalab.coderslab.com/v2/docs/).

According to the documentation, 4 endpoints have been made available:
 - `airport` - data about the airport,
 - `weather` - information about the recorded weather at the airport on a given day,
 - `aircraft` - data about aircraft
 - `flights` - data about departures from a given airport per day.

In order to retrieve information where the `airportId` parameter is required, I will used the list from the `airports.csv` file.

Token: `iKRsQ8vdqgT903o2vH1rsejOeQ0F7YC9TvutH6Wk`.

 ### Comments
 - API call limit is 1000/min

# import libraries
import requests
import pandas as pd
from datetime import datetime
from dateutil.relativedelta import relativedelta
import time
import os
from pandas import json_normalize

# API parameters
api_key = "API_KEY_TO_BE_REPLACED"
url = f'https://api-datalab.coderslab.com/api/v2'
headers = {"authorization": api_key}

# import airports.csv file and adaptation to further steps for downloading from endpoints
airports_df = pd.read_csv('./data/airports.csv')
airports_df
airport_ids = airports_df['origin_airport_id'].tolist()

# downloads airports data
def fetch_airport_data():
    airport_data = []
    for airport_id in airport_ids:
        response = requests.get(f"{url}/airport/{airport_id}", headers=headers)
        if response.status_code == 200:
            airport_data.append(response.json())
        time.sleep(0.1)  
    return pd.DataFrame.from_records(airport_data)

airport_df = fetch_airport_data()

airport_df

api_data_path = "./data/apidata"
isExist = os.path.exists(api_data_path)
print(isExist)

if not isExist:
    os.makedirs(api_data_path)

# saving the airport_df to the airport_list.csv
airport_df.to_csv(f'{api_data_path}/airport_list.csv', index=False)

# downloads weather data
def fetch_weather_data():
    start_date = datetime(2019, 1, 1)
    end_date = datetime(2020, 3, 31)
    weather_data = []

    for date in [start_date + relativedelta(months=x) for x in range((end_date.year - start_date.year) * 12 + end_date.month - start_date.month + 1)]:
        response = requests.get(f"{url}/airportWeather?date={date.strftime('%Y-%m')}", headers=headers)
        if response.status_code == 200:
            weather_data.extend(response.json())
        time.sleep(0.1)

    return pd.DataFrame(weather_data)

weather_df = fetch_weather_data()

airport_weather_df = weather_df
airport_weather_df

weather_df.to_csv(f'{api_data_path}/airport_weather.csv', index=False)

# downloads aircraft data
def fetch_aircraft_data():
    response = requests.get(f"{url}/aircraft", headers=headers)
    if response.status_code == 200:
        return pd.DataFrame(response.json())
    return pd.DataFrame()

aircraft_df = fetch_aircraft_data()
aircraft_df

aircraft_df.to_csv(f'{api_data_path}/aircraft.csv', index=False)

# downloads flights data
def fetch_flight_data(file_path, url, headers):
    # Wczytywanie pliku CSV
    airports_df = pd.read_csv(file_path)
    airport_ids = airports_df['origin_airport_id'].unique()
    
    # Defining a date range
    start_date = datetime(2019, 1, 1)
    end_date = datetime(2020, 3, 31)
    dates = [start_date + relativedelta(months=mon) for mon in range((end_date.year - start_date.year) * 12 + end_date.month - start_date.month + 1)]
    formatted_dates = [date.strftime('%Y-%m') for date in dates]
    
    # Downloading flight data
    flight_data = []
    for airport_id in airport_ids:
        no_data_found = False
        for date in formatted_dates:
            if no_data_found:
                break
            response = requests.get(f"{url}/flight?airportId={airport_id}&date={date}", headers=headers)
            if response.status_code == 200:
                f_data = response.json()
                flight_data.extend(f_data) 
                print(f"Data downloaded for airport ID {airport_id} for a month {date}.")
            elif response.status_code == 204:
                print(f"No data available for airport ID {airport_id} for a month {date}.")
                no_data_found = True
    
    return pd.DataFrame.from_records(flight_data)

# Path to the CSV file and API connection parameters
file_path = '.data/airports.csv'
url = f'https://api-datalab.coderslab.com/api/v2'
headers = {"authorization": api_key}

flight_df = fetch_flight_data(file_path, url, headers)
flight_df.shape

flight_df

flight_df.to_csv(f'{api_data_path}/flight.csv', index=False)
