==> 02_Initializing_the_Database.ipynb <==
 # Description of the notebook

In this notebook, I created a dedicated database with a table structure.

File with database structure: `database_schema.sql`

Database initialization done in notebook, not directly in Postgres.

import psycopg2
import os
from psycopg2 import connect
from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
import csv

# connection configuration
username = 'DB_ID'
password = 'DB_PASS'
host = 'localhost'
database = 'airlines'
default_database = 'postgres'
port = 5432

#Defining the `con` and `cursor` variables
con = psycopg2.connect(
        dbname=default_database,
        user=username,
        password=password,
        host=host,
        port=port
    )
cursor = con.cursor()
print("The connection to the database has been established.")

con.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)  
cursor = con.cursor()

# Rollback of any current transactions that may have resulted in an error
con.rollback()

# creating a database
cursor.execute("DROP DATABASE IF EXISTS airlines;")
cursor.execute("CREATE DATABASE airlines;")
print("The 'airlines' database has been created.")

#  Loading the database schema.sql file
file_path = './database_schema.sql'

with open(file_path, 'r') as file:
    sql_schema = file.read()

# splitting the contents of the `database schema.sql` file into smaller queries using `;`
queries = sql_schema.split(';')
queries = [query.strip() + ';' for query in queries if query.strip()]

# Execution of each query
for query in queries:
    if query:
        cursor.execute(query)

# Confirmation of all database operations, i.e. creation of tables
con.commit()

# Checking whether the databases have been added correctly

def check_if_table_exists(table_name):
    msg = f"I check if the table exists {table_name}"
    print(msg)

    query = f"select 1 from {table_name}"
    cursor.execute(query)
    print('OK!')


tables_to_test = [
    'aircraft',
    'airport_weather',
    'flight',
    'airport_list'
]

for table in tables_to_test:
    check_if_table_exists(table)

con.close()
