==> 03_Database_population.ipynb <==
 # notebook description
In this notebook, I appropriately adapted the data structures from the source files to a format compatible with `Postgres`, and then uploaded them to the server.

import psycopg2
from psycopg2 import connect
from sqlalchemy import create_engine
from sqlalchemy.engine import URL
import pandas as pd

username = 'DB_ID'
password = 'DB_PASS'
host = 'localhost'
database = 'airlines'
default_database = 'postgres'
port = 5432

url = URL.create(
    "postgresql+psycopg2",
    database = database,
    username=username,
    password=password,
    host=host,
    port=port    
    )
engine = create_engine(url)

def load_raw_data(file_name):
    df = pd.read_csv(file_name)
    
    df.columns = [col.lower() for col in df.columns]
    
    return df

# Loading individual files into frames
# File paths
api_data_path = "./data/apidata"
aircraft_file_path = f'{api_data_path}/aircraft.csv'
airport_list_file_path = f'{api_data_path}/airport_list.csv'
airport_weather_file_path = f'{api_data_path}/airport_weather.csv'
flight_file_path = f'{api_data_path}/flight.csv'

# Loading files
aircraft_df = load_raw_data(aircraft_file_path)
airport_list_df = load_raw_data(airport_list_file_path)
airport_weather_df = load_raw_data(airport_weather_file_path)
flight_df = load_raw_data(flight_file_path)

aircraft_df.head()

# Data export function
def export_table_to_db(df, table_name):
    print(f"Loading data into {table_name}...")

    df.reset_index(drop=True, inplace=True)
    df.index = df.index + 1

    try:
        global engine
        
        df.to_sql(name=table_name, con=engine, if_exists='replace', index=True, index_label='id', chunksize=500)
        print(f"Data successfully loaded into {table_name}.")
    except Exception as e:
        print(f"Failed to load data into {table_name}: {e}")

# Export first 5 rows for test
export_table_to_db(aircraft_df.head(5), 'aircraft')

# If the test succeeds, load the full data
export_table_to_db(aircraft_df, 'aircraft')

export_table_to_db(airport_weather_df.head(5), 'airport_weather')

export_table_to_db(airport_weather_df, 'airport_weather')

export_table_to_db(flight_df.head(5), 'flight')

export_table_to_db(flight_df, 'flight')

export_table_to_db(airport_list_df.head(5), 'airport_list')

export_table_to_db(airport_list_df, 'airport_list')
